+++ src/components/FormInitialText.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components"
+
+const FormInitialText = styled.div`
+    text-align: center;
+    margin-bottom: 2em;
+`
+export default FormInitialText
