+++ src/containers/signup-pf-p1/Signup_PF_P1.tsx
@@ -17,26 +17,7 @@ https://ant.design/components/overview/
 const Signup_PF_P1: React.FC = () => {
   return (
     <>
-      <Helmet>
-        <title>Sign Up</title>
-        <meta name="description" content="Description goes here." />
-      </Helmet>
-      <div className="content-container">
-        <div className="title">
-          <Title level={5}>SIGN UP</Title>
-          <Title level={3}>Registering as a Participating Family</Title>
-          <Paragraph>
-            Participating Families have early access to view events and are
-            eligible to attend free of charge. After creating an account, your
-            request will be reviewed by a member of our administration.
-          </Paragraph>
-          <Paragraph>Fields marked * are required.</Paragraph>
-        </div>
 
-        <Title level={5}>Account Owner Details</Title>
-
-      
-      </div>
     </>
   );
 };
